requirements.txt
# Data Extraction & Manipulation
statsbombpy==1.13.0
pandas==2.2.0

# Graph Database Connection
py2neo==2021.2.4
neo4j==5.17.0

# Network Analysis Metrics
networkx==3.2.1

# Visualization & UI
bokeh==3.3.4
Jinja2==3.1.3

# Environment Variable Management
python-dotenv==1.0.1